# documentation/openapi.yaml
openapi: 3.0.1
info:
  title: Musicman API
  description: Описание всех эндпоинтов сервиса "Musicman"
  contact:
    name: Анастасия Сашина
    email: sashina@cs.vsu.ru
  version: 1.0.6
servers:
  - url: http://localhost:8080/api/v1
    description: Generated server url
security:
  - Bearer: []
paths:
...
  "/songs":
    get:
      tags:
        - Song Controller
      summary: Выводит все песни по страницам
      operationId: getAllSongs
      parameters:
        - name: pageNumber
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
        - name: title
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: BY_TITLE
            enum:
              - BY_TITLE
              - BY_DURATION
        - name: isAsc
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: genres
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: singer
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Возвращает все песни по страницам
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongPagesDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
    post:
      tags:
        - Song Controller
      summary: Создание новой песни
      operationId: createSong
      requestBody:
        content:
          multipart/form-data:
            schema:
              "$ref": "#/components/schemas/CreateSongRequest"
        required: true
      responses:
        '201':
          description: Песня создана
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Отсутствует доступ
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Не существует исполнитель или жанр
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/library/{songId}":
    post:
      tags:
        - Audio Library Controller
      summary: Добавляет песню в аудиотеку
      operationId: addSong
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Песня добавлена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Песня не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '409':
          description: Песня уже в аудиотеке
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
    delete:
      tags:
        - Audio Library Controller
      summary: Удаляет песню из аудиотеки
      operationId: deleteSong_1
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Песня удалена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Песня не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '405':
          description: Песни нет в аудиотеке
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/songs/{id}":
    get:
      tags:
        - Song Controller
      summary: Выводит информацию о песне
      operationId: getSongById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Возвращает информацию о песне
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Песни не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
      security: []
    delete:
      tags:
        - Song Controller
      summary: Удаление песни
      operationId: deleteSong
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Песня удалена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Отсутствует доступ
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Песня не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/library":
    get:
      tags:
        - Audio Library Controller
      summary: Выводит все композиции в аудиотеке
      operationId: getAllSongs_1
      parameters:
        - name: pageNumber
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
        - name: title
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Возвращает все песни
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongPagesDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
  "/library/{songId}/exists":
    get:
      tags:
        - Audio Library Controller
      summary: Существует ли песня в аудиотеке
      operationId: songExists
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Возвращается результат
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongExistsDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Песня не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
components:
  schemas:
    ExceptionDTO:
      type: object
      properties:
        message:
          type: string
...
    CreateSongRequest:
      required:
        - duration
        - file
        - singerId
        - title
      type: object
      properties:
        title:
          maxLength: 255
          minLength: 0
          type: string
        duration:
          "$ref": "#/components/schemas/LocalTime"
        singerId:
          type: integer
          format: int64
        genreIds:
          type: array
          items:
            type: integer
            format: int64
        file:
          type: string
          format: binary
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    GenreDTO:
      type: object
      properties:
        genreId:
          type: integer
          format: int64
        name:
          type: string
...
    PagingInfoDTO:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    SingerDTO:
      type: object
      properties:
        singerId:
          type: integer
          format: int64
        fullName:
          type: string
        description:
          type: string
    SongDTO:
      type: object
      properties:
        songId:
          type: integer
          format: int64
        title:
          type: string
        duration:
          "$ref": "#/components/schemas/LocalTime"
        singer:
          "$ref": "#/components/schemas/SingerDTO"
        genres:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/GenreDTO"
    SongPagesDTO:
      type: object
      properties:
        paging:
          "$ref": "#/components/schemas/PagingInfoDTO"
        data:
          type: array
          items:
            "$ref": "#/components/schemas/SongDTO"
...
    SongExistsDTO:
      type: object
      properties:
        exists:
          type: boolean
...
  securitySchemes:
    Authorization:
      type: http
      scheme: Bearer
      bearerFormat: JWT